Add log parsing module
Comments are fairly informative, but implements a named 5-tuple for build
information.

Improves on the prototype by adding (tentative?) support for cancelled builds,
since these do not have a "rabbit2packer: ..." line but a "Cleanly cancelled
..." one.

Testing uses Doctest to extract examples from the docstrings -- but clearly
tests for equality using the __repr__ of the values, meaning that you'll
sometimes get obtuse messages like

          Expected "foo"
          But got 'foo'

When this happens, just switch to using the other style of quotes. (And maybe
add an issue upstream!)

Parsing uses regular expressions. :(