Fix formatting of logs
Currently using ansi2html to preserve the colours. However, having a wall of
green text is a bit of a strain on the eyes (of Alex) -- so I might just ditch
that. (Considering also that it's a terrible package.)

A future goal is to fix the CSS so that it'll scale properly to different sizes
of screen.