Refactor log getting/parsing to isolate effects
There is still some work to be done on this front -- it'd be good to do a
complete code review, but an initial stint with PyLint should be the first
step.

Also got rid of the doctests -- they clutter up the code, and I want to move
to Hypothesis' (QuickCheck inspired) property-based testing.